+++ Sources/MapboxMaps/Snapshot/Snapshotter.swift
@@ -26,6 +26,7 @@ internal protocol MapSnapshotterProtocol: StyleManagerProtocol {
                               pitch: NSNumber?) -> MapboxCoreMaps.CameraOptions
 
     func coordinateBoundsForCamera(forCamera camera: MapboxCoreMaps.CameraOptions) -> CoordinateBounds
+    func __tileCover(for options: MapboxCoreMaps.TileCoverOptions, cameraOptions: MapboxCoreMaps.CameraOptions?) -> [CanonicalTileID]
 }
 
 extension MapSnapshotter: MapSnapshotterProtocol {}
@@ -36,7 +37,7 @@ extension MapSnapshotter: MapSnapshotterProtocol {}
 /// You can configure the final result via ``MapSnapshotOptions`` upon construction time and take.
 public class Snapshotter: StyleManager {
 
-    /// Internal `MapboxCoreMaps.MBXMapSnapshotter` object that takes care of
+    /// Internal `MapboxCoreMaps.MBMMapSnapshotter` object that takes care of
     /// rendering a snapshot.
     internal let mapSnapshotter: MapSnapshotterProtocol
 
@@ -305,6 +306,17 @@ public class Snapshotter: StyleManager {
             padding: padding?.toMBXEdgeInsetsValue(),
             bearing: bearing?.NSNumber,
             pitch: pitch?.NSNumber))
+    }
+
+    /// Returns array of tile identifiers that cover current map camera.
+    ///
+    /// - Parameters:
+    ///  - options: Options for the tile cover method.
+    @_spi(Experimental)
+    public func tileCover(for options: TileCoverOptions) -> [CanonicalTileID] {
+        mapSnapshotter.__tileCover(
+            for: MapboxCoreMaps.TileCoverOptions(options),
+            cameraOptions: nil)
     }
 }